Some new entities imaged 2

+++ src/main/java/by/academy/it/task13/entity/PhotoSession.java
@@ -39,7 +39,7 @@ public class PhotoSession {
     private Integer horseCount;
 
     @Column(nullable = false)
-    private Double duration;
+    private Integer duration;
 
     @Column(nullable = false)
     private BigDecimal price;
